refactor: extract `useBalancesWithPool` hook
