+++ apps/dex/src/domains/bank/hooks/balances.ts
@@ -1,7 +1,9 @@
 import { Decimal } from "@cosmjs/math";
 import { useSigner } from "@sifchain/cosmos-connect";
 import { indexBy, prop } from "rambda";
+import { useMemo } from "react";
 import { useQuery } from "react-query";
+import { useLiquidityProviders } from "~/domains/clp";
 import { useDexEnvironment } from "~/domains/core/envs";
 import { useTokenRegistryQuery } from "~/domains/tokenRegistry";
 import { useSifSigningStargateClient } from "~/hooks/useSifStargateClient";
@@ -45,5 +47,50 @@ export const useAllBalances = () => {
       baseQuery.data === undefined
         ? undefined
         : indexBy(prop("denom"), baseQuery.data),
+  };
+};
+
+export const useBalancesWithPool = () => {
+  const { indexedByIBCDenom } = useTokenRegistryQuery();
+  const { data: liquidityProviders } = useLiquidityProviders();
+  const { data: balances } = useAllBalances();
+
+  const totalRowan = liquidityProviders?.pools.reduce(
+    (prev, curr) => prev.plus(curr.nativeAssetBalance),
+    Decimal.zero(18),
+  );
+
+  const denomSet = useMemo(
+    () =>
+      new Set([
+        ...(balances?.map((x) => x.denom) ?? []),
+        ...(liquidityProviders?.pools
+          .map((x) => x.liquidityProvider?.asset?.symbol as string)
+          .filter((x) => x !== undefined) ?? []),
+      ]),
+    [balances, liquidityProviders?.pools],
+  );
+
+  return {
+    balances: useMemo(
+      () =>
+        Array.from(denomSet).map((x) => {
+          const token = indexedByIBCDenom[x];
+          const balance = balances?.find((y) => y.denom === x);
+          const pool = liquidityProviders?.pools.find(
+            (y) => y.liquidityProvider?.asset?.symbol === x,
+          );
+          return {
+            denom: x,
+            symbol: token?.symbol,
+            displaySymbol: token?.displaySymbol,
+            network: token?.network,
+            amount: balance?.amount,
+            pool,
+          };
+        }),
+      [balances, denomSet, indexedByIBCDenom, liquidityProviders?.pools],
+    ),
+    totalRowan,
   };
 };
